Remove thrift encoder.
